+++ js/fetch-api.js
@@ -1,16 +1,25 @@
-import { createPointsMap } from './map-handler.js';
-
-const getData = (onError) => {
+/**
+ * Функция получения данных с сервера
+ * @param {Function} onSuccess - вызывается в случае успешной загрузки
+ * @param {Function} onError - вызывается в случае НЕуспешной загрузки
+ */
+const getData = (onSuccess, onError) => {
   fetch('https://23.javascript.pages.academy/keksobooking/data')
     .then((response) => response.json())
     .then((dataList) => {
-      createPointsMap(dataList);
+      onSuccess(dataList);
     })
     .catch(() => {
-      onError('Не удалось загрузить данные');
+      onError('Не удалось загрузить данные!');
     });
 };
 
+/**
+ * Функция отправки данных на сервер
+ * @param {Function} onSuccess - вызывается в случае успешной отправки
+ * @param {Function} onError - вызывается в случае НЕуспешной отправки
+ * @param {FormData} body - данные для отправки, введенные пользователем в формате FormData
+ */
 const sendData = (onSuccess, onError, body) => {
   fetch(
     'https://23.javascript.pages.academy/keksobooking',
