+++ js/util.js
@@ -1,7 +1,8 @@
-const isEscEvent = (evt) =>  evt.key === 'Escape' || evt.key === 'Esc';
-
-// const isEnterEvent = (evt) => {
-//   return evt.key === 'Enter';
-// };
+/**
+ * Функция проверки, что нажата клавиша Esc
+ * @param {Event} evt - событие
+ * @returns true если нажатая клавиша Esc и false в остальных случаях
+ */
+const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
 
 export { isEscEvent };
